+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes in 4.9
 - Fix month_date URL param for Attendance category tab link in TakeAttendance.php
 - Add Comments column to Grades in EditReportCardGrades.php, sponsored by ITS Japón school
 - CSS fix calendar below FlatSIS theme top menu in calendar-blue.css
+- Fix Grade Level for history grades in Transcripts.fnc.php
 
 Changes in 4.8.6
 ----------------